try add auto switch pure ipv6 network code ( need test )

+++ unity/Projects/Sources/xxnbsocketlib.cpp
@@ -55,7 +55,7 @@ extern "C" {
 
 	XXNBSOCKETLIB_API void xxnbs_set_address(XxNBSocket* nbs, char const* ip, int port) noexcept
 	{
-		return nbs->SetAddress(ip, port);
+		return nbs->SetAddress(ip, (uint16_t)port);
 	}
 
 	XXNBSOCKETLIB_API int xxnbs_connect(XxNBSocket* nbs, int sec, int usec) noexcept
